Add optional menu button in Main view.
This will be useful for touch-based browsers, where the context menu is
not accessible. Also, this will make it possible to set the Android
targetSdkVersion to 14, since the we'll no longer have to rely on the
legacy menu/overflow button from the system navigation bar. For
reference, see the "Say Goodbye to the Menu Button" post on the Android
Developers Blog:

http://android-developers.blogspot.com/2012/01/say-goodbye-to-menu-button.html