reloadTheme: preserve global ShockTherapyConfig
This fixes an invalid state, where the ShockTherapyMainView would hold
a reference to a stale ShockTherapyConfig instance. This only caused
problems when running under Sugar, due to the ShockTherapyConfig
implementation details for that environment. It's solved by making
reloadTheme preserve a global ShockTherapyConfig instance instead of
instantiating a new one.