Detect touch input and react appropriately.
This fixes menus to work correctly with Android 4.1.1 on the Nexus 7,
in both the WebView and in the Chrome web browser (which appear to
share the same internals). This browser is not detected as a "mobile"
browser by out current heuristics, so we'll detect it as a "touch"
browser instead.